Add Rhein-Neckar-Verkehr Motis
This patch adds a new Motis endpoint for GTFS feeds of
Rhein-Neckar-Verkehr GmbH, a local german operator that
runs trams and busses around Mannheim.

The Motis instance is run by me using Open Data by RNV,
with GTFS-RT currently relying on a testing access token
which expires at the end of the year. I hope that RNV will
see this and other uses of their data with Open Source tools
as valuable and keep providing me with a token.

The namespace "directions.nwex.de" is for a to be released
personal project of mine, for now I just use it to publish
various transport related apis and datasets.