Add `stopTimes` field to MOTIS
This patch adds a new `stopTimes` field to the options
for MOTIS apis. This field can specify a set of parameters
to be set as part of the request when using the `stoptimes`
api.

Currently the option includes `radius=200` for Transitous
and no values for RNV MOTIS.

While the MOTIS web interface defaults to a radius of 200
meters, setting this for all MOTIS endpoints unconditionally
causes wrong departures to be listed for deployments with
well maintained parent stations, such as the RNV MOTIS.

See https://github.com/derf/travelynx/issues/341